Assigned specific songs to specific waves and ambient moments.
Adjusted Jackson's Firestorm to be more optimized and fire less flames (but the flames are bigger now)

Fixed a bug that caused the Earth wall to disappear when the player pressed up against it.

Added health and Mana bars to the UI

Added a charge bar for the lightning charge shot.